+++ src/main/java/client/interaction/storage/ItemStorage.kt
@@ -2,4 +2,4 @@ package client.interaction.storage
 
 import client.inventory.ItemInventory
 
-class ItemStorage(slotMax: Short, var meso: Int) : ItemInventory(slotMax)
+class ItemStorage(slotMax: Byte, var meso: Int) : ItemInventory(slotMax)
