+++ src/main/java/client/inventory/ItemInventory.kt
@@ -2,7 +2,7 @@ package client.inventory
 
 import client.inventory.item.slots.ItemSlot
 
-open class ItemInventory(var slotMax: Short) {
+open class ItemInventory(var slotMax: Byte) {
 
     val items: MutableMap<Short, ItemSlot> = HashMap()
 }
